Add support for being notified when conn is idle
Enhance TcpConnection to provide methods to register callbacks for when
no data has been written to a connection or received on a connection for
a period of time. The methods take a time in milliseconds and a
Runnable callback that is run() in the event of the connection being
idle for more than the given number of milliseconds.

The functionality has been implemented in NettyTcpConnection using
Netty's IdleStateHandler. Unit tests for idle read and idle write have
also been added.